Finished the Countdown timer implementation.
Updated the user interface to display the correct information depending on the state of the timer.

Fixed resetButtonPressed() so that it will grab the new duration from the picker before resetting.

Added a stub for the did select row method (Homework: try and add a delegate protocol to update the user interface when the picker changes)